+++ input_handlers.py
@@ -259,3 +259,72 @@ def get_fixed32_input(current_value):
                     user_input += char  # Append only valid characters (digits or dots)
             except ValueError:
                 pass  # Ignore invalid inputs
+
+
+
+def select_from_list(prompt, current_option, list_options):
+    """
+    Displays a scrollable list of list_options for the user to choose from using a pad.
+    """
+    selected_index = list_options.index(current_option) if current_option in list_options else 0
+
+    height = min(len(list_options) + 5, curses.LINES - 2)
+    width = 60
+    start_y = (curses.LINES - height) // 2
+    start_x = (curses.COLS - width) // 2
+
+    list_win = curses.newwin(height, width, start_y, start_x)
+    list_win.bkgd(get_color("background"))
+    list_win.attrset(get_color("window_frame"))
+    list_win.keypad(True)
+
+    list_pad = curses.newpad(len(list_options) + 1, width - 8)
+    list_pad.bkgd(get_color("background"))
+
+    # Render header
+    list_win.clear()
+    list_win.border()
+    list_win.addstr(1, 2, prompt, get_color("settings_default", bold=True))
+
+    # Render options on the pad
+    for idx, color in enumerate(list_options):
+        if idx == selected_index:
+            list_pad.addstr(idx, 0, color.ljust(width - 8), get_color("settings_default", reverse=True))
+        else:
+            list_pad.addstr(idx, 0, color.ljust(width - 8), get_color("settings_default"))
+
+    # Initial refresh
+    list_win.refresh()
+    list_pad.refresh(0, 0,
+                    list_win.getbegyx()[0] + 3, list_win.getbegyx()[1] + 4,
+                    list_win.getbegyx()[0] + list_win.getmaxyx()[0] - 2, list_win.getbegyx()[1] + list_win.getmaxyx()[1] - 4)
+
+    while True:
+        key = list_win.getch()
+
+        if key == curses.KEY_UP:
+
+            if selected_index > 0:
+                selected_index -= 1
+
+        elif key == curses.KEY_DOWN:
+            if selected_index < len(list_options) - 1:
+                selected_index += 1
+
+        elif key == curses.KEY_RIGHT or key == ord('\n'):
+            return list_options[selected_index]
+        
+        elif key == curses.KEY_LEFT or key == 27:  # ESC key
+            return current_option
+
+        # Refresh the pad with updated selection and scroll offset
+        for idx, color in enumerate(list_options):
+            if idx == selected_index:
+                list_pad.addstr(idx, 0, color.ljust(width - 8), get_color("settings_default", reverse=True))
+            else:
+                list_pad.addstr(idx, 0, color.ljust(width - 8), get_color("settings_default"))
+
+        list_win.refresh()
+        list_pad.refresh(0, 0,
+                        list_win.getbegyx()[0] + 3, list_win.getbegyx()[1] + 4,
+                        list_win.getbegyx()[0] + list_win.getmaxyx()[0] - 2, list_win.getbegyx()[1] + list_win.getmaxyx()[1] - 4)
